Move signin components to core/auth Refactor navbar css
