+++ frontend-vue/src/components/core/auth/SignInCallback.vue
