--- frontend-vue/src/components/core/user.ts
@@ -1,3 +1,0 @@
-import { authStore } from "./SignIn.vue";
-
-export const { user } = storeToRefs(authStore);
