+++ frontend-vue/src/routes.ts
@@ -21,7 +21,7 @@ const DomainDetails = () => import("./pages/domain-details/DomainDetails.vue");
 const BoundedContextCanvas = () => import("./pages/bounded-context/BoundedContextCanvas.vue");
 const BoundedContextNamespaces = () => import("./pages/bounded-context/BoundedContextNamespaces.vue");
 const Analytics = () => import("./pages/analytics/Analytics.vue");
-const SignInCallback = () => import("./components/core/SignInCallback.vue");
+const SignInCallback = () => import("./components/core/auth/SignInCallback.vue");
 
 export default [
   {
